Handle all drum preparsing with a dedicated class
The issue with drums is that the type of drums being parsed isn't necessarily set in stone by a .ini. For .chart even, the "pro_drums" modifier has to get completely ignored (while .mid simply uses it as a guide). Tracking that state gets cumbersome externally, so I've internalized it with a class.