Restructure to support fallback to char[] on required encodings
Implements char[] versions of TXTReader, ChartReader, and IniReader.

Note: using them as default vs the original byte[] results in a 50% increase in scan time in sequential mode. And no, you still need to have a byte[] version for hashing, so forgoing the byte[] load is not an option. Not worth the performance loss NOR the extra memory usage.