Replace YargSong bool with a Cloneablestream
Instead of constantly passing around the state of whether the .sng uses YARG encryption, solely determine that state through the file's header tag. The existence of "YARGSONG" means the obvious.

Holding the stream in the object has the added benefit of ensuring the file is still valid for access for any of the listings that may use (unless a drive error occurs, of course).

+ Reverts cache version