Pass ProcessParams by `ref` instead of `in`
`in` is supposed to be the equivalent of a `readonly ref`. It should only be used in specific cases where functions will make no modifications, under any circumstance, to any value of the parameter.

The problem here is that this was not the case. Some functions, during parse, would actively modify the drum enum value of the ParseSettings variable within a ProcessParams parameter through a local reference. This worked fine when ParseSettings was a class (despite the bad practice), but now that it's a struct, that local reference became a local copy. Any modification would not exist past the end of the function's execution, which leads to an error being thrown when attempting to load a drum track with an original drum enum value of `Unknown`.

Not only does using `ref`` fix this issue, it more clearly defines the intended usage of the parameter at all call sites.