Convert FixedArray into a small hierarchy of fixed array types [Breaking change]
All types contains pointers and lengths. What differs is how the memory is allocated. One is the new memorymappedArray which uses the backing of a memory mapped file while the other is for new empty allocations.

Note: FixedArray's base class will function simply as an alias. It simply holds the length & pointer values without modification or disposal. It's only the subclasses that handle creation and disposal.