Convert SngMask to a struct
Utilizing an unsafe fixed buffer simplifies the basic functionality of the type.
Firstly, it removes the need to have some tracker count how many references to the fixed buffer exist. This means SngMask (and thus SngFile) no longer have to be disposable.

Secondly, making it a struct provides inherent cloning behavior. We don't need to add a special function to mimic it when the values will be memcopied by default.