Utilize A Hierarchy Of FixedArray Types To Enhance Memory Usage (Part 1) (#182)
* Convert FixedArray into a small hierarchy of fixed array types [Breaking change]

All types contains pointers and lengths. What differs is how the memory is allocated. One is the new memorymappedArray which uses the backing of a memory mapped file while the other is for new empty allocations.

Note: FixedArray's base class will function simply as an alias. It simply holds the length & pointer values without modification or disposal. It's only the subclasses that handle creation and disposal.

* Add Length variable to AbridgedFileInfo

This will be necessary for some MemoryMapped functionality later

* Add MemoryMappedArray Load() overload for AbridgedFileInfo

* Move Length variable to separate Abridged structure

* Utilize FixedArray hierarchy to enhance memory usage across scanning architecture

Requires some more stringent Disposable pattens for certain types to ensure all data is cleared after their applicable scopes.

Later benefits: more freedom to use direct pointers in situation where bounds are pre-guaranteed to dodge certain overheads.

* Add specialization for loading UTF-16 TextReaders

In the cases where the file uses UTF-16, and where the endianness of the file matches the host system's memory layout, there no need to realloc and copy the file. A simple cast of the original data suffices.

* AbridgedFileInfo syntax fixes