IniModifier Refactor (#191)
* Ini modifier handling refactor

Simplified the structure to move the type checking to outside functions. The important piece of the type checking is that it should only be a debug process. It should not be necessary outside of release builds. So, we turn on the type checks only within `#DEBUG` regions.

Through converting the type to a struct, we can directly cast its internal buffer to the desired type. No need for a union to handle that job. Extra benefit: we can convert most of the TryGet()s in IniSection to a generic method.

+ Moved all the .chart modifiers to the main IniHandler dictionary. This allows for the type checking.

* Switch ExtractSngString param from `ref` to `in`

* Remove unused `partial` tag on IniModifier

* Fix "Offset" typo